# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /nest/paste-drafts/{membershipId}/{kind}:
    parameters:
      - { name: membershipId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: kind, in: path, required: true, schema: { type: string, enum: [groceries, chores] } }
    get:
      operationId: getNestPasteDraft
      responses:
        "200":
          description: Account-private draft, or a revisioned empty slot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NestPasteDraft" }
        "401": { description: Sign in required }
        "403": { description: Membership unavailable }
    put:
      operationId: putNestPasteDraft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NestPasteDraftInput" }
      responses:
        "200":
          description: Draft checkpoint or discard committed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NestPasteDraft" }
        "400": { description: Invalid draft }
        "401": { description: Sign in required }
        "403": { description: Membership unavailable }
        "409": { description: Draft changed; load latest before continuing }
  /nest/paste-drafts/{membershipId}/{kind}/save:
    parameters:
      - { name: membershipId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: kind, in: path, required: true, schema: { type: string, enum: [groceries, chores] } }
    post:
      operationId: saveNestPasteDraftRow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NestPasteRowInput" }
      responses:
        "200":
          description: Record and remaining draft committed atomically
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NestPasteDraft" }
        "400": { description: Invalid reviewed row }
        "401": { description: Sign in required }
        "403": { description: Membership unavailable }
        "409": { description: Draft changed; no item sent }
  /nest/attachments:
    post:
      operationId: uploadNestJournalAttachment
      summary: Stage a private journal attachment
      parameters:
        - in: query
          name: name
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "201":
          description: Private staged file
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NestJournalAttachment" }
        "400": { description: Invalid or empty file }
        "401": { description: Sign in required }
        "403": { description: Household membership required }
        "413": { description: File exceeds 10 MB }
        "415": { description: Unsupported file type }
        "429": { description: Too many unsaved files }
  /nest/attachments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getNestJournalAttachment
      summary: Read a file after checking the current journal visibility
      responses:
        "200":
          description: Authorized file bytes; only raster images may display inline
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { description: Sign in required }
        "404": { description: File unavailable }
    delete:
      operationId: discardNestJournalAttachment
      summary: Idempotently discard an own staged upload; saved attachments are untouched
      responses:
        "204": { description: Draft cleanup finished }
        "401": { description: Sign in required }
        "404": { description: Invalid ID or no household }
  /nest/attachments/{id}/download:
    get:
      operationId: downloadNestJournalAttachment
      summary: Download an authorized attachment, including raster images
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Authorized file download
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { description: Sign in required }
        "404": { description: File unavailable }
...
components:
  schemas:
    NestPasteRow:
      type: object
      additionalProperties: false
      required: [id, title, quantity, assignedTo, cadence, dueDate]
      properties:
        id: { type: string, minLength: 1, maxLength: 100 }
        title: { type: string, maxLength: 1000 }
        quantity: { type: string, maxLength: 1000 }
        assignedTo: { type: string, maxLength: 100 }
        cadence: { type: string, maxLength: 100 }
        dueDate: { type: string, maxLength: 100 }
        isPaid: { type: boolean }
        payAmount: { type: string, maxLength: 100 }
    NestPasteSession:
      type: object
      additionalProperties: false
      required: [text, drafts, reviewing, savedCount, showErrors, error, uncertainId, finished]
      properties:
        text: { type: string, maxLength: 20000 }
        drafts:
          type: array
          maxItems: 200
          items: { $ref: "#/components/schemas/NestPasteRow" }
        reviewing: { type: boolean }
        savedCount: { type: integer, minimum: 0, maximum: 100000 }
        showErrors: { type: boolean }
        error: { type: string, maxLength: 2000 }
        uncertainId: { type: [string, "null"], maxLength: 100 }
        finished: { type: boolean }
    NestPasteDraft:
      type: object
      required: [revision, session]
      properties:
        revision: { type: [string, "null"], format: uuid }
        session:
          oneOf:
            - { $ref: "#/components/schemas/NestPasteSession" }
            - { type: "null" }
    NestPasteDraftInput:
      type: object
      additionalProperties: false
      required: [revision, session]
      properties:
        revision: { type: [string, "null"], format: uuid }
        session:
          oneOf:
            - { $ref: "#/components/schemas/NestPasteSession" }
            - { type: "null" }
    NestPasteRowInput:
      type: object
      additionalProperties: false
      required: [revision, rowId]
      properties:
        revision: { type: string, format: uuid }
        rowId: { type: string, minLength: 1, maxLength: 100 }
...
    NestJournalAttachment:
      type: object
      additionalProperties: false
      required: [id, name, contentType, size]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        contentType: { type: string }
        size: { type: integer, minimum: 1, maximum: 10485760 }
...