# lib/api-zod/src/generated/types/nestPasteRowInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NestPasteRowInput {
  revision: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  rowId: string;
}